commit 15.06.2025

+++ db/migrations/20250608121514_add_type_to_quest.js
@@ -1,11 +1,17 @@
 export async function up(knex) {
-  return knex.schema.alterTable('quests', (table) => {
-    table.string('type').defaultTo('daily');
-  });
+  const hasColumn = await knex.schema.hasColumn('quests', 'type');
+  if (!hasColumn) {
+    return knex.schema.alterTable('quests', (table) => {
+      table.string('type').defaultTo('daily');
+    });
+  }
 }
 
 export async function down(knex) {
-  return knex.schema.alterTable('quests', (table) => {
-    table.dropColumn('type');
-  });
+  const hasColumn = await knex.schema.hasColumn('quests', 'type');
+  if (hasColumn) {
+    return knex.schema.alterTable('quests', (table) => {
+      table.dropColumn('type');
+    });
+  }
 }
